implement handle custom error

+++ src/storage/group/groupCreate.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AppError } from "../../utils/AppError";
 import { GROUP_COLLECTION } from "../storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
 
@@ -6,6 +7,11 @@ export async function groupCreate(newGroup: string) {
     try {
         const storedGroups = await groupsGetAll();
 
+        const groupAreadyExists = storedGroups.includes(newGroup);
+
+        if (groupAreadyExists) {
+            throw new AppError('Ja existe um grupo cadastrado com esse nome.');
+        }
         const storage = JSON.stringify([...storedGroups, newGroup]);
         await AsyncStorage.setItem(GROUP_COLLECTION, storage);
     } catch (error) {
